+++ src/hooks/goldens/zsh-menu-command-not-found.zsh
@@ -187,16 +187,16 @@ cd+() {
   forward "$@"
 }
 
+cdr() {
+  __dx_jump_mode recents "${1:-}"
+}
+
 cdf() {
   __dx_jump_mode frecents "${1:-}"
 }
 
 z() {
   cdf "$@"
-}
-
-cdr() {
-  __dx_jump_mode recents "${1:-}"
 }
 
 _dx_complete_paths() {